Heavily modified IsotonicLogisticRegression.py but let it generate same results as before.
Implemented test_with_simulation, rendering most of the previous tests obsolete. Also fixed several bugs with this test.
Implemented feat_pvalue_method='auto' for autosetting method to get p-values (spearman-R for regression and Mann-Whitney-U for classification)
Changed spearman_r_pvalue_... to feat_pvalue_...
Changed X[...] to X1[...] in self._split
Let get_feature_importances take stat_test as an optional param to return the given statistics (-inf to inf or -1 to 1), p-values (0 to 1), and trends (-1, 0, +1).
Please be aware that the statistic, p-values, and trends of the odds_spearmanr test may not be accurate from empirical test.
Implemented and used self.get_average_y() the regression-equivalent of the self.get_odds_offset() for classification.
Added and implemented the optional param increasings (a vector of booleans, a string, or None) to the self.fit method to specify if each column is increasing or decreasing.
Implemented mannwhitneyu-based test.
Let self.fit_transform pass kwargs to the self.fit method instead of not passing them.